requirements.txt
datasets
langchain
langchain-mongodb
langchain-openai
pymongo
pandas
python-dotenv
streamlit
langchain_community
